Create README.md
This is the very first project I made at Rutgers University in the class Introduction to Computer Science. Some classes were given to us such as CFGUI.java, while others were used and made as part of our assignment grade.I created the classes AIPlayer, HumanPlayer and Board which helped implement the gameplay of the program, and ConnectFour.java which designed the overall GUI by using what was given to us.